# openapi.json
{"openapi":"3.0.1","info":{"title":"BookStore REST API","description":"REST API for BookStore modular monolith application. Provides endpoints for managing products, shopping cart, and orders. Built with Spring Boot and Spring Modulith.","license":{"name":"Apache 2.0","url":"https://www.apache.org/licenses/LICENSE-2.0"},"version":"1.0.0"},"servers":[{"url":"http://localhost:8080","description":"Local development server"}],"tags":[{"name":"Cart","description":"Shopping cart management API"},{"name":"Orders","description":"Order management API"},{"name":"Products","description":"Product catalog management API"}],"paths":{"/api/cart/items/{code}":{"put":{"tags":["Cart"],"summary":"Update item quantity","description":"Updates the quantity of an item in the cart","operationId":"updateItemQuantity","parameters":[{"name":"code","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateQuantityRequest"}}},"required":true},"responses":{"400":{"description":"Invalid quantity","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CartDto"}}}},"200":{"description":"Item quantity updated successfully","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CartDto"}}}},"404":{"description":"Item not found in cart","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CartDto"}}}}}}},"/api/orders":{"get":{"tags":["Orders"],"summary":"List orders","description":"Retrieves a paginated list of orders","operationId":"listOrders","parameters":[{"name":"page","in":"query","description":"Page number (1-based)","required":false,"schema":{"type":"integer","format":"int32","default":1}},{"name":"pageSize","in":"query","description":"Page size","required":false,"schema":{"type":"integer","format":"int32","default":10}}],"responses":{"503":{"description":"Orders service unavailable","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PagedResult"}}}},"200":{"description":"Orders retrieved successfully","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PagedResult"}}}}}},"post":{"tags":["Orders"],"summary":"Create order","description":"Creates a new order from the provided order details","operationId":"createOrder","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateOrderRequest"}}},"required":true},"responses":{"201":{"description":"Order created successfully","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CreateOrderResponse"}}}},"503":{"description":"Orders service unavailable","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CreateOrderResponse"}}}},"400":{"description":"Invalid order data","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CreateOrderResponse"}}}}}}},"/api/cart/items":{"post":{"tags":["Cart"],"summary":"Add item to cart","description":"Adds a product to the shopping cart or updates quantity if already exists","operationId":"addItem","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AddToCartRequest"}}},"required":true},"responses":{"400":{"description":"Invalid request data","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CartDto"}}}},"404":{"description":"Product not found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CartDto"}}}},"201":{"description":"Item added to cart successfully","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CartDto"}}}}}}},"/api/products":{"get":{"tags":["Products"],"summary":"Get paginated products","description":"Retrieves a paginated list of products from the catalog","operationId":"getProducts","parameters":[{"name":"page","in":"query","description":"Page number (1-based)","required":false,"schema":{"type":"integer","format":"int32","default":1},"example":1}],"responses":{"200":{"description":"Successfully retrieved product list","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PagedResult"}}}}}}},"/api/products/{code}":{"get":{"tags":["Products"],"summary":"Get product by code","description":"Retrieves a specific product by its unique code","operationId":"getProductByCode","parameters":[{"name":"code","in":"path","description":"Product code","required":true,"schema":{"type":"string"},"example":"P100"}],"responses":{"404":{"description":"Product not found"},"200":{"description":"Successfully retrieved product","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProductDto"}}}}}}},"/api/orders/{orderNumber}":{"get":{"tags":["Orders"],"summary":"Get order details","description":"Retrieves detailed information about a specific order","operationId":"getOrder","parameters":[{"name":"orderNumber","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"404":{"description":"Order not found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OrderDto"}}}},"503":{"description":"Orders service unavailable","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OrderDto"}}}},"200":{"description":"Order retrieved successfully","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OrderDto"}}}}}}},"/api/cart":{"get":{"tags":["Cart"],"summary":"Get cart","description":"Retrieves the current shopping cart contents","operationId":"getCart","responses":{"200":{"description":"Cart retrieved successfully","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CartDto"}}}}}},"delete":{"tags":["Cart"],"summary":"Clear cart","description":"Removes all items from the shopping cart","operationId":"clearCart","responses":{"204":{"description":"Cart cleared successfully"}}}}},"components":{"schemas":{"UpdateQuantityRequest":{"required":["quantity"],"type":"object","properties":{"quantity":{"minimum":1,"type":"integer","description":"New quantity","format":"int32","example":3}},"description":"Request to update cart item quantity"},"CartDto":{"required":["items"],"type":"object","properties":{"items":{"type":"array","description":"Cart items","items":{"$ref":"#/components/schemas/CartItemDto"}},"totalAmount":{"type":"number","description":"Total amount","example":99.99},"itemCount":{"type":"integer","description":"Number of items in cart","format":"int32","example":3}},"description":"Shopping cart with items and total"},"CartItemDto":{"type":"object","properties":{"code":{"type":"string","description":"Product code","example":"P001"},"name":{"type":"string","description":"Product name","example":"Spring Boot in Action"},"price":{"type":"number","description":"Unit price","example":29.99},"quantity":{"minimum":1,"type":"integer","description":"Quantity","format":"int32","example":2},"subtotal":{"type":"number","description":"Subtotal (price * quantity)","example":59.98}},"description":"Shopping cart line item"},"CreateOrderRequest":{"required":["customer","deliveryAddress","item"],"type":"object","properties":{"customer":{"$ref":"#/components/schemas/Customer"},"deliveryAddress":{"type":"string","description":"Delivery address","example":"742 Evergreen Terrace, Springfield"},"item":{"$ref":"#/components/schemas/OrderItem"}},"description":"Request to create a new order"},"Customer":{"required":["email","name","phone"],"type":"object","properties":{"name":{"type":"string","description":"Customer full name","example":"John Doe"},"email":{"type":"string","description":"Customer email address","example":"john.doe@example.com"},"phone":{"type":"string","description":"Customer phone number","example":"+1-555-123-4567"}},"description":"Customer information for order"},"OrderItem":{"required":["code","name","price","quantity"],"type":"object","properties":{"code":{"type":"string","description":"Product code","example":"P100"},"name":{"type":"string","description":"Product name","example":"The Hunger Games"},"price":{"type":"number","description":"Product price","example":34.0},"quantity":{"minimum":1,"type":"integer","description":"Order quantity","format":"int32","example":2}},"description":"Order line item representing a product in the order"},"CreateOrderResponse":{"required":["orderNumber"],"type":"object","properties":{"orderNumber":{"type":"string","description":"Unique order number","example":"ORD-2025-001234"}},"description":"Response after creating an order"},"AddToCartRequest":{"required":["code","quantity"],"type":"object","properties":{"code":{"type":"string","description":"Product code","example":"P001"},"quantity":{"minimum":1,"type":"integer","description":"Quantity to add","format":"int32","example":1}},"description":"Request to add item to cart"},"PagedResult":{"required":["data","hasNext","hasPrevious","isFirst","isLast","pageNumber","totalElements","totalPages"],"type":"object","properties":{"data":{"type":"array","description":"List of items for the current page","items":{"type":"object","description":"List of items for the current page"}},"totalElements":{"type":"integer","description":"Total number of elements across all pages","format":"int64","example":100},"pageNumber":{"minimum":1,"type":"integer","description":"Current page number (1-based)","format":"int32","example":1},"totalPages":{"minimum":0,"type":"integer","description":"Total number of pages","format":"int32","example":10},"isFirst":{"type":"boolean","description":"Whether this is the first page","example":true},"isLast":{"type":"boolean","description":"Whether this is the last page","example":false},"hasNext":{"type":"boolean","description":"Whether there is a next page available","example":true},"hasPrevious":{"type":"boolean","description":"Whether there is a previous page available","example":false}},"description":"Paginated response containing data and pagination metadata"},"ProductDto":{"required":["code","name","price"],"type":"object","properties":{"code":{"type":"string","description":"Unique product code","example":"P100"},"name":{"type":"string","description":"Product name","example":"The Hunger Games"},"description":{"type":"string","description":"Product description","example":"Winning will make you famous. Losing means certain death..."},"imageUrl":{"type":"string","description":"Product image URL","example":"https://images.sivalabs.in/products/the-hunger-games.jpg"},"price":{"type":"number","description":"Product price","example":34.0}},"description":"Product information"},"OrderView":{"required":["customer","orderNumber","status"],"type":"object","properties":{"orderNumber":{"type":"string","description":"Unique order number","example":"ORD-2025-001234"},"status":{"type":"string","description":"Order status","example":"NEW","enum":["NEW","IN_PROCESS","DELIVERED","CANCELLED","ERROR"]},"customer":{"$ref":"#/components/schemas/Customer"}},"description":"Summary view of an order"},"OrderDto":{"required":["createdAt","customer","deliveryAddress","item","orderNumber","status"],"type":"object","properties":{"orderNumber":{"type":"string","description":"Unique order number","example":"ORD-2025-001234"},"item":{"$ref":"#/components/schemas/OrderItem"},"customer":{"$ref":"#/components/schemas/Customer"},"deliveryAddress":{"type":"string","description":"Delivery address","example":"742 Evergreen Terrace, Springfield"},"status":{"type":"string","description":"Order status","example":"NEW","enum":["NEW","IN_PROCESS","DELIVERED","CANCELLED","ERROR"]},"createdAt":{"type":"string","description":"Order creation timestamp","format":"date-time"},"totalAmount":{"type":"number","description":"Total order amount (price × quantity)","readOnly":true,"example":68.0}},"description":"Complete order details"}}}}